Use CFFI in API mode for geopmdpy and geopmpy
