+++ geopmdpy/geopmdpy/shmem.py
@@ -6,12 +6,6 @@
 from . import gffi
 from . import error
 
-gffi.gffi.cdef("""
-
-    int geopm_shmem_create_prof(const char *shm_key, size_t size, int pid, int uid, int gid);
-    int geopm_shmem_path_prof(const char *shm_key, int pid, int uid, size_t shm_path_max, char *shm_path);
-
-""")
 _dl = gffi.get_dl_geopmd()
 
 def create_prof(shm_key, size, pid, uid, gid):
