+++ geopmpy/geopmpy/policy_store.py
@@ -5,32 +5,15 @@
 
 
 import math
-from geopmdpy.gffi import gffi
-from geopmdpy.gffi import get_dl_geopm
+from geopmdpy import gffi
 from geopmdpy import error
 
-gffi.cdef("""
-int geopm_policystore_connect(const char *data_path);
-
-int geopm_policystore_disconnect();
-
-int geopm_policystore_get_best(const char* agent_name, const char* profile_name,
-                               size_t max_policy_vals, double* policy_vals);
-
-int geopm_policystore_set_best(const char* agent_name, const char* profile_name,
-                               size_t num_policy_vals, const double* policy_vals);
-
-int geopm_policystore_set_default(const char* agent_name,
-                                  size_t num_policy_vals, const double* policy_vals);
-""")
-_dl = get_dl_geopm()
-
-
-def is_implemented():
-    """Check whether the optional PolicyStore feature is implemented in this
-    deployment of GEOPM.
-    """
-    return hasattr(_dl, 'geopm_policystore_connect')
+_dl = gffi.get_dl_geopm()
+if not hasattr(_dl, 'geopm_policystore_connect'):
+    raise ImportError('geopmpy.policy_store cannot be imported because the installed '
+                      'libgeopm does not include the PolicyStore feature. '
+                      'Rebuild libgeopm with the --enable-beta configuration flag '
+                      'then reinstall geopmpy.')
 
 
 def connect(database_path):
@@ -40,10 +23,9 @@ def connect(database_path):
     Args:
         database_path (str): Path to the database.
     """
-    global gffi
     global _dl
 
-    database_path_cstr = gffi.new("char[]", database_path.encode())
+    database_path_cstr = gffi.libgeopm_ffi.new("char[]", database_path.encode())
     err = _dl.geopm_policystore_connect(database_path_cstr)
     if err < 0:
         raise RuntimeError('geopm_policystore_connect() failed: {}'.format(error.message(err)))
@@ -71,13 +53,12 @@ def get_best(agent_name, profile_name):
     Returns:
         list[float]: Best known policy for the profile and agent.
     """
-    global gffi
     global _dl
 
-    agent_name_cstr = gffi.new("char[]", agent_name.encode())
-    profile_name_cstr = gffi.new("char[]", profile_name.encode())
+    agent_name_cstr = gffi.libgeopm_ffi.new("char[]", agent_name.encode())
+    profile_name_cstr = gffi.libgeopm_ffi.new("char[]", profile_name.encode())
     policy_max = 1024
-    policy_array = gffi.new("double[]", policy_max)
+    policy_array = gffi.libgeopm_ffi.new("double[]", policy_max)
     err = _dl.geopm_policystore_get_best(agent_name_cstr, profile_name_cstr,
                                          policy_max, policy_array)
     if err < 0:
@@ -94,12 +75,11 @@ def set_best(agent_name, profile_name, policy):
         profile_name (str): Name of the profile.
         policy (list[float]): New policy to use.
     """
-    global gffi
     global _dl
 
-    agent_name_cstr = gffi.new("char[]", agent_name.encode())
-    profile_name_cstr = gffi.new("char[]", profile_name.encode())
-    policy_array = gffi.new("double[]", policy)
+    agent_name_cstr = gffi.libgeopm_ffi.new("char[]", agent_name.encode())
+    profile_name_cstr = gffi.libgeopm_ffi.new("char[]", profile_name.encode())
+    policy_array = gffi.libgeopm_ffi.new("double[]", policy)
     err = _dl.geopm_policystore_set_best(agent_name_cstr, profile_name_cstr,
                                          len(policy), policy_array)
     if err < 0:
@@ -113,11 +93,10 @@ def set_default(agent_name, policy):
         agent_name (str): Name of the agent.
         policy (list[float]): Default policy to use with the agent.
     """
-    global gffi
     global _dl
 
-    agent_name_cstr = gffi.new("char[]", agent_name.encode())
-    policy_array = gffi.new("double[]", policy)
+    agent_name_cstr = gffi.libgeopm_ffi.new("char[]", agent_name.encode())
+    policy_array = gffi.libgeopm_ffi.new("double[]", policy)
     err = _dl.geopm_policystore_set_default(agent_name_cstr, len(policy), policy_array)
     if err < 0:
         raise RuntimeError('geopm_policystore_set_default() failed: {}'.format(error.message(err)))
