+++ geopmpy/pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools>=59.6", "setuptools-scm>=6.4.2" ]
+requires = ["setuptools>=59.6", "setuptools-scm>=6.4.2", "cffi>=1.14.5"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools]
